+++ services/explorer-ui/src/styles/global.css
@@ -15,7 +15,7 @@
     --blue: hsl(227, 92%, 62%); /* #456cf7 */
     --green: hsl(158, 60%, 45%); /* #31A67B */
     --red: hsl(7, 100%, 70%); /* #FF7764 */
-    --yellow: hsl(49, 80%, 75%); /* #FFE471 */
+    --yellow: hsl(49, 90%, 70%); /* #FFE471 */
     --purple-dark: hsl(262, 36%, 21%); /* #302249 */
     --purple-light: hsl(251, 45%, 56%); /* #715EC2 */
